Refactor namings
